backend/requirements.txt
Flask==3.1.1
Flask-Cors==6.0.1
bcrypt==4.3.0
psycopg[binary]==3.2.13
psycopg-pool==3.2.6
python-dotenv==1.1.1
